+++ projet/lib/src/capteurs_distance.cpp
@@ -17,20 +17,20 @@ void CapteursDistance::init()
 	DDRA &= ~(_BV(_BROCHE_TO_PIN(BROCHE_CAPTEUR_DISTANCE_DROIT)));
 }
 
-uint16_t CapteursDistance::getDistanceGauche()
+uint8_t CapteursDistance::getDistanceGauche()
 {
     return canToDistance(can_.lecture(_BROCHE_TO_PIN(BROCHE_CAPTEUR_DISTANCE_GAUCHE)));
 }
 
-uint16_t CapteursDistance::getDistanceDroit()
+uint8_t CapteursDistance::getDistanceDroit()
 {
     return canToDistance(can_.lecture(_BROCHE_TO_PIN(BROCHE_CAPTEUR_DISTANCE_DROIT)));
 }
 
-uint16_t CapteursDistance::canToDistance(const uint16_t canVal)
+uint8_t CapteursDistance::canToDistance(const uint8_t canVal)
 {
     // voir util/python/interpolationCapteurDistance
-    uint16_t dist = (6787.0 / ((double)canVal - 3.0)) - 4.0;
+    uint8_t dist = (6787.0 / ((double)canVal - 3.0)) - 4.0;
 
     // validation
     dist = (CAPTEUR_DISTANCE_MIN <= dist && dist <= CAPTEUR_DISTANCE_MAX)
